requirements.txt
attrs==16.1.0
funcparserlib==0.3.6
future==0.15.2
pendulum==0.5.5
py==1.4.31
pytest==3.0.2
python-dateutil==2.5.3
pytz==2016.6.1
six==1.10.0
tabulate==0.7.5
tzlocal==1.2.2
